web_upload/requirements.txt
Flask==2.1.1
PyYAML==5.4.1
SQLAlchemy==1.4.27
mysql-connector-python==8.0.26
sqlalchemy==1.4.27